+++ DynAIkonTrap/camera.py
@@ -122,6 +122,16 @@ class Camera:
             self._output = read_from
         logger.debug("Camera started")
 
+    def capture_still(self, filename: str, fmt='image/jpeg',):
+        """Captures a still image on the video port, can be called while the camera is recording.
+        Insipred by picamera example: https://picamera.readthedocs.io/en/release-1.13/recipes2.html#capturing-images-whilst-recording
+        Args: 
+            filename (str): filename to save capture to, jpg.
+        """
+        logger.info("Still image requested...")
+        self._camera.capture(filename, format=fmt, use_video_port=True)
+        logger.info(f"Sill saved to: {filename}")
+
     def get(self) -> Frame:
         """Retrieve the next frame from the camera
 
@@ -148,3 +158,4 @@ class Camera:
 
     def close(self):
         self._camera.stop_recording()
+        self._camera.close()